+++ src/Metrics/CarbonSink.h
@@ -21,8 +21,8 @@ public:
   CarbonSink(std::string Host, uint16_t const Port)
       : CarbonConnection(std::move(Host), Port){};
   void reportMetric(InternalMetric &MetricToBeReported) override;
-  LogTo getType() override { return LogTo::CARBON; };
-  bool isHealthy() override;
+  LogTo getType() const override { return LogTo::CARBON; };
+  bool isHealthy() const override;
 
 private:
   Carbon::Connection CarbonConnection;