+++ src/Status/StatusReporterBase.h
@@ -15,7 +15,7 @@
 #include "logger.h"
 #include <asio.hpp>
 #include <chrono>
-#include <mutex>
+#include <shared_mutex>
 
 namespace flatbuffers {
 class DetachedBuffer;
@@ -24,6 +24,7 @@ class DetachedBuffer;
 namespace Status {
 
 using JsonGeneratorFuncType = std::function<void(nlohmann::json &)>;
+using StatusGetterFuncType = std::function<Status::JobStatusInfo()>;
 
 class StatusReporterBase {
 public:
@@ -44,27 +45,12 @@ public:
 
   virtual ~StatusReporterBase() = default;
 
-  /// \brief Set the slow changing information to report.
-  ///
-  /// \param NewInfo The new information to report
-  virtual void updateStatusInfo(JobStatusInfo const &NewInfo);
-
   virtual void useAlternativeStatusTopic(std::string const &AltTopicName);
 
   virtual void revertToDefaultStatusTopic();
 
-  /// \brief Update the stop time to be reported.
-  ///
-  /// \param StopTime The new stop time.
-  virtual void updateStopTime(time_point StopTime);
-
   /// \brief Get the currently reported stop time.
-  virtual time_point getStopTime();
-
-  /// \brief Clear out the current information.
-  ///
-  /// Used when a file has finished writing.
-  virtual void resetStatusInfo();
+  virtual time_point getStopTime() const;
 
   /// \brief Generate a FlatBuffer serialised report.
   ///
@@ -77,8 +63,13 @@ public:
 
   virtual void
   setJSONMetaDataGenerator(JsonGeneratorFuncType GeneratorFunction) {
-    std::lock_guard Lock(StatusMutex);
+    std::unique_lock const Lock(StatusMutex);
     JSONGenerator = GeneratorFunction;
+  }
+
+  void setStatusGetter(StatusGetterFuncType NewStatusGetter) {
+    std::unique_lock const Lock(StatusMutex);
+    StatusGetter = NewStatusGetter;
   }
 
 protected:
@@ -87,14 +78,18 @@ protected:
 
 private:
   virtual void postReportStatusActions(){};
-  JobStatusInfo Status{};
-  mutable std::mutex StatusMutex;
+  mutable std::shared_mutex StatusMutex;
   std::shared_ptr<Kafka::Producer> Producer;
   std::unique_ptr<Kafka::ProducerTopic> StatusProducerTopic;
   std::unique_ptr<Kafka::ProducerTopic> AltStatusProducerTopic;
   bool UsingAlternativeStatusTopic{false};
   ApplicationStatusInfo const StaticStatusInformation;
-  std::function<void(nlohmann::json &JSONNode)> JSONGenerator;
+  JsonGeneratorFuncType JSONGenerator{[](auto) {
+    throw std::runtime_error("JSONGenerator(): Not set/implemented.");
+  }};
+  StatusGetterFuncType StatusGetter{[]() -> Status::JobStatusInfo {
+    throw std::runtime_error("StatusGetter(): Not set/implemented.");
+  }};
 };
 
 } // namespace Status